chat_service/requirements.txt
Django==5.2.1
celery==5.5.3
django-cors-headers==4.7.0
redis==4.6.0
djangorestframework==3.16.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
channels==4.2.2
daphne==4.2.0
channels_redis==4.2.1
pika==1.3.2
djangorestframework_simplejwt==5.5.0
PyJWT==2.9.0
